+++ examples/client/getBlocksAbort.ts
@@ -1,4 +1,4 @@
-import { createConcordiumClient, HexString } from '@concordium/node-sdk';
+import { createConcordiumClient, ArrivedBlockInfo } from '@concordium/node-sdk';
 import { credentials } from '@grpc/grpc-js';
 
 import meow from 'meow';
@@ -8,21 +8,13 @@ const cli = meow(
   Usage
     $ yarn ts-node <path-to-this-file> [options]
 
-  Required
-    --height,     The block height to get blocks from
-
   Options
-    --help,         Displays this message
+    --help,     -h  Displays this message
     --endpoint, -e  Specify endpoint of the form "address:port", defaults to localhost:20000
 `,
     {
         importMeta: import.meta,
         flags: {
-            height: {
-                type: 'number',
-                alias: 'h',
-                isRequired: true,
-            },
             endpoint: {
                 type: 'string',
                 alias: 'e',
@@ -40,15 +32,24 @@ const client = createConcordiumClient(
 );
 
 /**
- * Get a list of live blocks at a given absolute height.
+ * Returns a stream of blocks that is iterable. The following code will receive
+ * a single block and then abort:
  */
 
 (async () => {
-    const blocks: HexString[] = await client.getBlocksAtHeight(
-        BigInt(cli.flags.height)
+    // Create abort controller and block stream
+    const ac = new AbortController();
+    const blockStream: AsyncIterable<ArrivedBlockInfo> = client.getBlocks(
+        ac.signal
     );
 
-    for (const block of blocks) {
-        console.log(block);
+    // Only get one item then break
+    for await (const block of blockStream) {
+        console.log('Arrived block height:', block.height);
+        console.log('Arrived block hash:', block.hash, '\n');
+        break;
     }
+
+    // Closes the stream
+    ac.abort();
 })();