+++ examples/client/getInstanceState.ts
@@ -1,4 +1,7 @@
-import { BakerId, createConcordiumClient } from '@concordium/node-sdk';
+import {
+    createConcordiumClient,
+    InstanceStateKVPair,
+} from '@concordium/node-sdk';
 import { credentials } from '@grpc/grpc-js';
 
 import meow from 'meow';
@@ -6,16 +9,24 @@ import meow from 'meow';
 const cli = meow(
     `
   Usage
-    $ yarn run-example <path-to-this-file> [options]
+    $ yarn ts-node <path-to-this-file> [options]
+
+  Required:
+    --contract, -c  The index of the contract to query info about
 
   Options
-    --help,         Displays this message
+    --help,     -h  Displays this message
     --block,    -b  A block to query from, defaults to last final block
     --endpoint, -e  Specify endpoint of the form "address:port", defaults to localhost:20000
 `,
     {
         importMeta: import.meta,
         flags: {
+            contract: {
+                type: 'number',
+                alias: 'c',
+                isRequired: true,
+            },
             block: {
                 type: 'string',
                 alias: 'b',
@@ -38,20 +49,25 @@ const client = createConcordiumClient(
 );
 
 /**
- * Retrieves a stream of ID's for registered bakers on the network at a specific
- * block. If a blockhash is not supplied it will pick the latest finalized
- * block. An optional abort signal can also be provided that closes the stream.
+ * Get the exact state of a specific contract instance, streamed as a list of
+ * hex string key-value pairs. If a blockhash is not supplied it will pick the
+ * latest finalized block. An optional abortSignal can also be provided that
+ * closes the stream.
 
  * Note: A stream can be collected to a list with the streamToList function.
  */
 
 (async () => {
-    const bakerIds: AsyncIterable<BakerId> = client.getBakerList(
+    const contractAddress = {
+        index: BigInt(cli.flags.contract),
+        subindex: 0n,
+    };
+    const states: AsyncIterable<InstanceStateKVPair> = client.getInstanceState(
+        contractAddress,
         cli.flags.block
     );
 
-    console.log('List of BakerID at the specified block:');
-    for await (const bakerId of bakerIds) {
-        console.log(bakerId);
+    for await (const state of states) {
+        console.dir(state, { depth: null, colors: true });
     }
 })();