+++ examples/client/getNodeInfo.ts
@@ -1,4 +1,4 @@
-import { BakerId, createConcordiumClient } from '@concordium/node-sdk';
+import { createConcordiumClient, NodeInfo } from '@concordium/node-sdk';
 import { credentials } from '@grpc/grpc-js';
 
 import meow from 'meow';
@@ -6,10 +6,10 @@ import meow from 'meow';
 const cli = meow(
     `
   Usage
-    $ yarn run-example <path-to-this-file> [options]
+    $ yarn ts-node <path-to-this-file> [options]
 
   Options
-    --help,         Displays this message
+    --help,     -h  Displays this message
     --block,    -b  A block to query from, defaults to last final block
     --endpoint, -e  Specify endpoint of the form "address:port", defaults to localhost:20000
 `,
@@ -38,20 +38,19 @@ const client = createConcordiumClient(
 );
 
 /**
- * Retrieves a stream of ID's for registered bakers on the network at a specific
- * block. If a blockhash is not supplied it will pick the latest finalized
- * block. An optional abort signal can also be provided that closes the stream.
+ * Get information about the node.
 
- * Note: A stream can be collected to a list with the streamToList function.
+ * The `NodeInfo` includes information of:
+ *  - Meta information, such as the, version of the node, type of the node, uptime
+ *    and the local time of the node.
+ *  - NetworkInfo, which yields data such as the node id, packets sent/received,
+ *    average bytes per second sent/received.
+ *  - ConsensusStatus. The `ConsensusStatus` returned depends on if the node supports
+ *    the protocol on chain and whether the node is configured as a baker or not.
  */
 
 (async () => {
-    const bakerIds: AsyncIterable<BakerId> = client.getBakerList(
-        cli.flags.block
-    );
+    const nodeInfo: NodeInfo = await client.getNodeInfo();
 
-    console.log('List of BakerID at the specified block:');
-    for await (const bakerId of bakerIds) {
-        console.log(bakerId);
-    }
+    console.dir(nodeInfo, { depth: null, colors: true });
 })();