+++ packages/common/src/util.ts
@@ -5,6 +5,7 @@ import {
     IpAddressString,
     ReleaseSchedule,
 } from './types';
+import { Buffer } from 'buffer/';
 
 /**
  * Replaces a number in a JSON string with the same number as a
@@ -138,6 +139,26 @@ export function countSignatures(
 }
 
 /**
+ * Converts a wasm module to a smart contract schema.
+ * @param wasm the wasm module as a Buffer
+ * @returns the smart contract schema as a Buffer
+ */
+export function wasmToSchema(wasm: Buffer): Buffer {
+    const wasmModule = new WebAssembly.Module(wasm);
+    const sections = WebAssembly.Module.customSections(
+        wasmModule,
+        'concordium-schema'
+    );
+    if (sections.length === 1) {
+        return Buffer.from(sections[0]);
+    } else if (sections.length === 0) {
+        throw Error('WASM-Module contains no schema!');
+    } else {
+        throw Error('Invalid WASM-Module retrieved!');
+    }
+}
+
+/**
  * Convert a Date to seconds since epoch.
  */
 export function secondsSinceEpoch(date: Date): bigint {
@@ -252,3 +273,7 @@ export const makeDynamicFunction =
     <T, R>(fun: (a: T[]) => R) =>
     (input: T | T[]): R =>
         fun(Array.isArray(input) ? input : [input]);
+
+export function isDefined<T>(v?: T): v is T {
+    return v !== undefined;
+}